+++ tests/tt_eager/python_api_testing/non_working_unit_tests/wormhole/test_eltwise_heaviside.py
@@ -7,10 +7,6 @@ import sys
 from loguru import logger
 import numpy as np
 import random
-
-f = f"{Path(__file__).parent}"
-sys.path.append(f"{f}/../../../../..")
-
 import pytest
 import torch
 import tt_lib as ttl
