+++ tests/tt_eager/python_api_testing/non_working_unit_tests/wormhole/test_eltwise_logaddexp2.py
@@ -6,10 +6,6 @@ from pathlib import Path
 import sys
 from loguru import logger
 import random
-
-f = f"{Path(__file__).parent}"
-sys.path.append(f"{f}/../../../../..")
-
 import pytest
 import torch
 import tt_lib as ttl
