+++ tests/tt_eager/python_api_testing/non_working_unit_tests/wormhole/test_eltwise_signbit.py
@@ -8,10 +8,6 @@ import sys
 import time
 import os
 from loguru import logger
-
-f = f"{Path(__file__).parent}"
-sys.path.append(f"{f}/../../../../..")
-
 import pytest
 import torch
