+++ tests/tt_eager/python_api_testing/non_working_unit_tests/wormhole/test_layernorm_noweights.py
@@ -8,10 +8,6 @@ import sys
 import time
 import os
 from loguru import logger
-
-f = f"{Path(__file__).parent}"
-sys.path.append(f"{f}/../../../../..")
-
 import pytest
 import torch
 
@@ -19,7 +15,7 @@ import tt_lib as ttl
 from tests.tt_eager.python_api_testing.sweep_tests import pytorch_ops
 from tests.tt_eager.python_api_testing.sweep_tests.comparison_funcs import comp_pcc, comp_equal
 from tests.tt_eager.python_api_testing.sweep_tests.common import is_wormhole_b0, skip_for_wormhole_b0
-from tests.tt_eager.python_api_testing.sweep_tests.tt_lib_ops import layernorm as tt_layernorm
+from tests.tt_eager.python_api_testing.sweep_tests.tt_lib_ops import layernorm_noweights as tt_layernorm
 
 
 def run_layernorm_tests(
@@ -45,21 +41,19 @@ def run_layernorm_tests(
     print(in_mem_config)
 
     x = torch.Tensor(size=input_shape[0]).uniform_(-10, 10)
-    y = torch.Tensor(size=input_shape[1]).uniform_(-10, 10)
-    z = torch.Tensor(size=input_shape[2]).uniform_(-10, 10)
+    # y = torch.Tensor(size=input_shape[1]).uniform_(-10, 10)
+    # z = torch.Tensor(size=input_shape[2]).uniform_(-10, 10)
 
     x_ref = x.detach().clone()
-    y_ref = y.detach().clone()
-    z_ref = z.detach().clone()
+    # y_ref = y.detach().clone()
+    # z_ref = z.detach().clone()
 
     # compute ref value --------------------------
-    ref_value = pytorch_ops.layernorm(x_ref, y_ref, z_ref)
+    ref_value = pytorch_ops.layernorm_noweights(x_ref)
 
     # compute tt value ---------------------------
     tt_result = tt_layernorm(
         x=x,
-        y=y,
-        z=z,
         device=device,
         device_id=0,
         dtype=dtype,
@@ -71,9 +65,7 @@ def run_layernorm_tests(
     # compare tt and golden outputs -------------
     success, pcc_value = comp_equal(ref_value, tt_result)
     logger.debug(pcc_value)
-
     assert success
-
 
 test_sweep_args=[
     (
