+++ tests/tt_eager/python_api_testing/non_working_unit_tests/wormhole/test_rpow.py
@@ -10,9 +10,6 @@ import os
 from loguru import logger
 from models.utility_functions import torch2tt_tensor, tt2torch_tensor
 import random
-f = f"{Path(__file__).parent}"
-sys.path.append(f"{f}/../../../../..")
-
 import pytest
 import torch
 
@@ -28,12 +25,12 @@ from tests.tt_eager.python_api_testing.sweep_tests.comparison_funcs import comp_
 from tests.tt_eager.python_api_testing.sweep_tests.common import is_wormhole_b0, skip_for_wormhole_b0
 
 
-def ref_rdiv(x, factor):
-    return x/factor
+def ref_rpow(x, factor):
+    return torch.pow(x, factor)
 
 
 
-def run_rdiv_tests(input_shape, dtype, dlayout, in_mem_config, out_mem_config, data_seed, device):
+def run_rpow_tests(input_shape, dtype, dlayout, in_mem_config, out_mem_config, data_seed, device):
     torch.manual_seed(data_seed)
 
     # Initialize the device
@@ -46,7 +43,7 @@ def run_rdiv_tests(input_shape, dtype, dlayout, in_mem_config, out_mem_config, d
         for nrepeat in range(0, 100):
             x = torch.Tensor(size=(N, C, H, W)).uniform_(-100, 100)
             x_ref = x
-            factor = random.uniform(0,2)
+            factor = random.randint(1,100)
             if dlayout == ttl.tensor.Layout.TILE:
                 x = tilize_to_list(x)
             else:
@@ -70,8 +67,8 @@ def run_rdiv_tests(input_shape, dtype, dlayout, in_mem_config, out_mem_config, d
                     in_mem_config,
                     )
 
-            logger.info("Running rdiv test")
-            ttz = tensor.rdiv(ttx, factor, output_mem_config=out_mem_config)
+            logger.info("Running rpow test")
+            ttz = tensor.rpow(ttx, factor, output_mem_config=out_mem_config)
 
 
             logger.info("Done")
@@ -88,14 +85,14 @@ def run_rdiv_tests(input_shape, dtype, dlayout, in_mem_config, out_mem_config, d
             output = tt2torch_tensor(ttz)
 
             # get referent value
-            ref_value = ref_rdiv(factor, x_ref)
+            ref_value = ref_rpow(factor, x_ref)
 
 
             # compare tt and golden outputs
             success, pcc_value = comp_pcc(output, ref_value)
             logger.debug(pcc_value)
             logger.debug(success)
-            #assert success
+            assert success
 
 
 test_sweep_args=[
@@ -112,7 +109,7 @@ test_sweep_args=[
     ),
 )
 
-def test_rdiv(
+def test_rpow(
     input_shape, dtype, dlayout, in_mem_config, out_mem_config, data_seed, device
 ):
-    run_rdiv_tests(input_shape, dtype, dlayout, in_mem_config, out_mem_config, data_seed, device)
+    run_rpow_tests(input_shape, dtype, dlayout, in_mem_config, out_mem_config, data_seed, device)
